Minor CI fixes
